add hello world tests.
