+++ docker-compose.yml
@@ -8,7 +8,7 @@ services:
       - "8080:8080"
     volumes:
       - .:/go/src/work # マウントディレクトリ指定
-    command: go run main.go
+    command: go run httpServer.go
   db:
     image: mysql:5.7
     container_name: mysql_host
