+++ hello/hello.go
@@ -1,5 +1,5 @@
 package hello
 
 func GetHello(s string) string {
-  return s + "Hello, World!"
+  return s + " Hello, World!"
 }
