--- hellomain.go
@@ -1,11 +1,0 @@
-package main
-
-import (
-    "fmt"
-    "./hello"
-)
-  
-func main() {
-  s := hello.GetHello("HI! Guest! ")
-  fmt.Println(s)
-}
